+++ apps/ui-dev/constants/main-pages.js
@@ -1,11 +1,23 @@
 export default [
   "Introduction",
   {
-    link: "/pages/introduction/getting-started.html",
+    link: "/pages/introduction/getting-started/",
     title: "Getting Started",
   },
   {
-    link: "/pages/introduction/install.html",
+    link: "/pages/introduction/install/",
     title: "Install",
+  },
+  {
+    link: "/pages/introduction/core/",
+    title: "Core Concept",
+  },
+  {
+    link: "/pages/introduction/develop/",
+    title: "Development",
+  },
+  {
+    link: "/pages/introduction/contribute/",
+    title: "Contribute",
   },
 ];