+++ apps/ui-dev/pages/introduction/Test.jsx
@@ -1,0 +1,3 @@
+export function Test() {
+  return <h1>Test</h1>;
+}