+++ apps/ui-dev/utils/Diagram.d.ts
@@ -1,0 +1,10 @@
+import { UIElement } from "@elf-framework/sapa";
+
+interface DiagramProps {
+  graph: string;
+  diagram: string;
+}
+
+export class Diagram extends UIElement {
+  props: DiagramProps;
+}