+++ apps/ui-dev/utils/Diagram.jsx
@@ -1,15 +1,22 @@
 import { htmlToVNode, useEffect, useState } from "@elf-framework/sapa";
-import mermaid from "mermaid";
 
-export function Diagram({ diagram }) {
+export function Diagram({ diagram, graph }) {
   const [content, setContent] = useState("");
 
   useEffect(() => {
-    if (!content) {
-      mermaid.mermaidAPI.render("id1", diagram, (svg) => {
-        setContent(htmlToVNode(svg));
-      });
-    }
+
+    (async () => {
+      const mermaid = (await import("mermaid")).default;
+
+      if (!content) {
+        mermaid.mermaidAPI.render("id1", graph || diagram, (svg) => {
+          setContent(htmlToVNode(svg));
+        });
+      }
+
+    })();
+
+
   }, [content, diagram]);
 
   return <div>{content}</div>;